fix 1 : replaced seperator to separator fix 2 : put enum and interface to respective folders in the breadcrumbs component fix 3 : use of "first" in ngFor instead of non-efficient logic
