+++ libs/ui/src/lib/breadcrumbs/breadcrumbs.component.ts
@@ -1,4 +1,7 @@
 import { Component, Input } from '@angular/core';
+import { BreadcrumbDisplay } from './enums/breadcrumb-display.enum';
+import { BreadcrumbSeparator } from './enums/breadcrumb-separator.enum';
+import { Breadcrumb } from './interfaces/breadcrumb.interface';
 
 /**
  * UI Breadcrumbs Component
@@ -10,36 +13,10 @@ import { Component, Input } from '@angular/core';
 })
 export class BreadcrumbsComponent {
   @Input() breadcrumbs: Breadcrumb[] = [];
-  @Input() seperator: BreadcrumbSeperator = BreadcrumbSeperator.CHEVRON;
+  @Input() separator: BreadcrumbSeparator = BreadcrumbSeparator.CHEVRON;
   @Input() display: BreadcrumbDisplay = BreadcrumbDisplay.SIMPLE;
 
   //In order to be able to use enumerations in html
   BreadcrumbsTypes = BreadcrumbDisplay;
-  BreadcrumbsSeperators = BreadcrumbSeperator;
-}
-
-/**
- * Enumeration of possible display styles
- */
-export enum BreadcrumbDisplay {
-  SIMPLE,
-  CONTAINED,
-  FULL,
-}
-
-/**
- * Enumeration of possible seperators styles
- */
-export enum BreadcrumbSeperator {
-  CHEVRON,
-  SLASH,
-}
-
-/** Interface of breadcrumb */
-export interface Breadcrumb {
-  alias?: string;
-  uri: string;
-  text?: string;
-  key?: string;
-  queryParams?: any;
+  BreadcrumbsSeparators = BreadcrumbSeparator;
 }
