+++ libs/ui/src/lib/breadcrumbs/enums/breadcrumb-display.enum.ts
@@ -1,0 +1,8 @@
+/**
+ * Enumeration of possible display styles
+ */
+export enum BreadcrumbDisplay {
+  SIMPLE,
+  CONTAINED,
+  FULL,
+}
