+++ libs/ui/src/lib/breadcrumbs/enums/breadcrumb-separator.enum.ts
@@ -1,0 +1,7 @@
+/**
+ * Enumeration of possible separators styles
+ */
+export enum BreadcrumbSeparator {
+  CHEVRON,
+  SLASH,
+}
