+++ libs/ui/src/lib/breadcrumbs/interfaces/breadcrumb.interface.ts
@@ -1,0 +1,8 @@
+/** Interface of breadcrumb */
+export interface Breadcrumb {
+  alias?: string;
+  uri: string;
+  text?: string;
+  key?: string;
+  queryParams?: any;
+}
